Fixes security analysis result deps cycles
